Updated the Map
The map and the doors with a few props are in the game. Everything
should be in their place scaled correctly. All that's left to make the
game playable is to have the enemies. If there's time after the enemies
are created, I'm going to add more props to give the ship more life.